pyproject.toml
[tool.poetry]
name = "image-url-classifier"
version = "0.1.0"
description = "A simple url image classifier"
authors = ["James Hirst <hirst.jj@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^3.0.3"
requests = "^2.31.0"
Pillow = "^10.4.0"
torch = "^2.4.0"
torchvision = "^0.19.0"
validators = "^0.33.0"
numpy = "1.26.4"

[tool.poetry.dev-dependencies]
black = "^23.0.0"
isort = "^5.8.0"
mypy = "^1.3.0"
pytest = "^7.0.0"

[tool.poetry.group.dev.dependencies]
types-pillow = "^10.2.0.20240520"
types-requests = "^2.32.0.20240712" 

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
